+++ js/const.js
@@ -1,0 +1,35 @@
+export const DESCRIPTION = [
+  'Красивый закат на берегу моря',
+  'Музей современного искусства',
+  'Театр оперы и балета',
+  'Центральная улица города',
+  'Гроза в деревне'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const NAMES = [
+  'Антон',
+  'Александр',
+  'Артур',
+  'Зорик',
+  'Влад',
+  'Таня',
+  'Батор'
+];
+
+// мин и макс значения id,лайков,аватаров и комментариев;количество объектов в итоговом массиве
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+export const MIN_AVATAR = 1;
+export const MAX_AVATAR = 6;
+export const MIN_COMMENTS = 0;
+export const MAX_COMMENTS = 30;
+export const SIMILAR_OBJECTS_COUNT = 25;
